+++ lib/src/connection.dart
@@ -99,6 +99,10 @@ abstract class Connection {
 
   final Completer<void> _onInitialPeerSettingsReceived = Completer<void>();
 
+  final StreamController<int> _pingReceived = StreamController<int>();
+
+  final StreamController<void> _frameReceived = StreamController<void>();
+
   /// Future which completes when the first SETTINGS frame is received from
   /// the peer.
   Future<void> get onInitialPeerSettingsReceived =>
@@ -179,12 +183,12 @@ abstract class Connection {
     // Setup handlers.
     _settingsHandler = SettingsHandler(_hpackContext.encoder, _frameWriter,
         acknowledgedSettings, peerSettings);
-    _pingHandler = PingHandler(_frameWriter);
+    _pingHandler = PingHandler(_frameWriter, _pingReceived);
 
     var settings = _decodeSettings(settingsObject);
 
     // Do the initial settings handshake (possibly with pushes disabled).
-    _settingsHandler.changeSettings(settings).catchError((error) {
+    _settingsHandler.changeSettings(settings).catchError((Object error) {
       // TODO: The [error] can contain sensitive information we now expose via
       // a [Goaway] frame. We should somehow ensure we're only sending useful
       // but non-sensitive information.
@@ -267,15 +271,15 @@ abstract class Connection {
   }
 
   /// Pings the remote peer (can e.g. be used for measuring latency).
-  Future ping() {
+  Future<void> ping() {
     return _pingHandler.ping().catchError((e, s) {
-      return Future.error(
+      return Future<void>.error(
           TransportException('The connection has been terminated.'));
     }, test: (e) => e is TerminatedException);
   }
 
   /// Finishes this connection.
-  Future finish() {
+  Future<void> finish() {
     _finishing(active: true);
 
     // TODO: There is probably more we need to wait for.
@@ -284,8 +288,8 @@ abstract class Connection {
   }
 
   /// Terminates this connection forcefully.
-  Future terminate() {
-    return _terminate(ErrorCode.NO_ERROR);
+  Future<void> terminate([int? errorCode]) {
+    return _terminate(errorCode ?? ErrorCode.NO_ERROR);
   }
 
   void _activeStateHandler(bool isActive) =>
@@ -339,6 +343,9 @@ abstract class Connection {
     } else if (frame is PushPromiseFrame) {
       frame.decodedHeaders =
           _hpackContext.decoder.decode(frame.headerBlockFragment);
+    }
+    if (_frameReceived.hasListener) {
+      _frameReceived.add(null);
     }
 
     // Handle the frame as either a connection or a stream frame.
@@ -434,16 +441,15 @@ abstract class Connection {
       _settingsHandler.terminate(exception);
 
       return Future.wait([cancelFuture, closeFuture])
-          .catchError((_) => const []);
+          .catchError((_) => const <void>[]);
     }
-    return Future.value();
+    return Future<void>.value();
   }
 }
 
 class ClientConnection extends Connection implements ClientTransportConnection {
-  ClientConnection._(Stream<List<int>> incoming, StreamSink<List<int>> outgoing,
-      Settings settings)
-      : super(incoming, outgoing, settings, isClientConnection: true);
+  ClientConnection._(super.incoming, super.outgoing, super.settings)
+      : super(isClientConnection: true);
 
   factory ClientConnection(Stream<List<int>> incoming,
       StreamSink<List<int>> outgoing, ClientSettings clientSettings) {
@@ -473,12 +479,17 @@ class ClientConnection extends Connection implements ClientTransportConnection {
     }
     return hStream;
   }
+
+  @override
+  Stream<int> get onPingReceived => _pingReceived.stream;
+
+  @override
+  Stream<void> get onFrameReceived => _frameReceived.stream;
 }
 
 class ServerConnection extends Connection implements ServerTransportConnection {
-  ServerConnection._(Stream<List<int>> incoming, StreamSink<List<int>> outgoing,
-      Settings settings)
-      : super(incoming, outgoing, settings, isClientConnection: false);
+  ServerConnection._(super.incoming, super.outgoing, super.settings)
+      : super(isClientConnection: false);
 
   factory ServerConnection(Stream<List<int>> incoming,
       StreamSink<List<int>> outgoing, ServerSettings serverSettings) {
@@ -489,4 +500,10 @@ class ServerConnection extends Connection implements ServerTransportConnection {
   @override
   Stream<ServerTransportStream> get incomingStreams =>
       _streams.incomingStreams.cast<ServerTransportStream>();
+
+  @override
+  Stream<int> get onPingReceived => _pingReceived.stream;
+
+  @override
+  Stream<void> get onFrameReceived => _frameReceived.stream;
 }