+++ lib/src/streams/stream_handler.dart
@@ -283,7 +283,7 @@ class StreamHandler extends Object with TerminatableMixin, ClosableMixin {
             'not in "idle" state.');
       }
 
-      var sameDirection = (nextStreamId + remoteStreamId) % 2 == 0;
+      var sameDirection = (nextStreamId + remoteStreamId).isEven;
       assert(!sameDirection);
 
       lastRemoteStreamId = remoteStreamId;
@@ -343,7 +343,7 @@ class StreamHandler extends Object with TerminatableMixin, ClosableMixin {
     // NOTE: We are not interested whether the streams were normally finished
     // or abnormally terminated. Therefore we use 'catchError((_) {})'!
     var streamDone = [streamQueueIn.done, streamQueueOut.done];
-    Future.wait(streamDone).catchError((_) => const []).whenComplete(() {
+    Future.wait(streamDone).catchError((_) => const <void>[]).whenComplete(() {
       _cleanupClosedStream(stream);
     });
 
@@ -593,11 +593,26 @@ class StreamHandler extends Object with TerminatableMixin, ClosableMixin {
         } else if (frame is PushPromiseFrame) {
           throw ProtocolException('Cannot push on a non-existent stream '
               '(stream ${frame.header.streamId} does not exist)');
+        } else if (frame is DataFrame) {
+          // http/2 spec:
+          //     However, after sending the RST_STREAM, the sending endpoint
+          //     MUST be prepared to receive and process additional frames sent
+          //     on the stream that might have been sent by the peer prior to
+          //     the arrival of the RST_STREAM.
+          // and:
+          //     A receiver that receives a flow-controlled frame MUST always
+          //     account for its contribution against the connection
+          //     flow-control window, unless the receiver treats this as a
+          //     connection error (Section 5.4.1). This is necessary even if the
+          //     frame is in error. The sender counts the frame toward the
+          //     flow-control window, but if the receiver does not, the
+          //     flow-control window at the sender and receiver can become
+          //     different.
+          incomingQueue.processIgnoredDataFrame(frame);
+          // Still respond with an error, as the stream is closed.
+          throw _throwStreamClosedException(frame.header.streamId);
         } else {
-          throw StreamClosedException(
-              frame.header.streamId,
-              'No open stream found and was not a headers frame opening a '
-              'new stream.');
+          throw _throwStreamClosedException(frame.header.streamId);
         }
       } else {
         if (frame is HeadersFrame) {
@@ -868,4 +883,11 @@ class StreamHandler extends Object with TerminatableMixin, ClosableMixin {
     var id = stream.id;
     return (isServer && id.isEven) || (!isServer && id.isOdd);
   }
+
+  static Exception _throwStreamClosedException(int streamId) =>
+      StreamClosedException(
+        streamId,
+        'No open stream found and was not a headers frame opening a '
+        'new stream.',
+      );
 }